backend/requirements.txt
# ==============================================================================
# AssistantAudit v2 — Backend Dependencies (pinned for reproducible builds)
# ==============================================================================

# --- Framework web ---
fastapi==0.135.1
uvicorn[standard]==0.42.0

# --- Base de données ---
SQLAlchemy==2.0.48
alembic==1.18.4
psycopg2-binary>=2.9.0  # installed only in production (PostgreSQL)

# --- Chiffrement & Certificats ---
cryptography==46.0.6

# --- WebSocket ---
websockets==16.0

# --- Validation & Sérialisation ---
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0

# --- Authentification ---
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.22

# --- YAML (référentiels) ---
PyYAML==6.0.3

# --- Sécurité XML (scan Nmap) ---
defusedxml==0.7.1

# --- Collecte SSH / WinRM ---
paramiko==4.0.0
pywinrm==0.5.0
python-dateutil==2.9.0.post0

# --- Audit Active Directory (LDAP) ---
ldap3==2.9.1

# --- Logging & Config ---
python-dotenv==1.2.2

# --- HTTP client (pour futures intégrations) ---
httpx==0.28.1

# --- Dev & Tests ---
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1

# --- Logging Structure ---
python-json-logger==4.0.0

# --- Monitoring & Metrics ---
prometheus-client==0.24.1
sentry-sdk==2.55.0

# --- Génération de rapports ---
WeasyPrint==63.1
python-docx==1.1.2
Jinja2==3.1.6
